racket: fix formulation of tnum operator precision
The old formulation was wrong because the call to "symbolics" in
tnum->symbolic would include the symbolics from the num as well as the
newly created constant.  This moved variables which should have been
universally quantified into the existential quantifier.

Fixing this uncovered that the intended property is correct, but
too strong.  It asserts that the tnum operation is as precise as the
corresponding operation on sets of bitvectors. This rejects operators
like tnum-add, which are not maximally precise on all sets of inputs,
but is as precise as one can possibly get with tnums.

This commit fixes the problem with a formulation in Rosette that is
equivalent to the natural definition of optimality that quantifies over
functions of tnums, with a Lean proof of equivalence.

For optimality proofs, decrease the bitwidth to avoid killing Z3.