racket/verifier: weaken verifier invariants for mov32
The old invariants were strong enough for scalar_min_max_or, and weak
enough to be maintained by the instructions I've verified so far.

To validate the invariants, I implemented them as a C function in the
kernel and modified the verifier to panic if the invariants were ever
violated. I discovered one such violation in the `test_progs` selftests,
due to the mov32 BPF instruction.

The mov32 instruction bypasses adjust_scalar_min_max_vals in the verifier,
and therefore bypasses __update_reg_bounds, __reg_deduce_bounds,
and __reg_bound_offset, which maintain the invariants by refining the
{32,64}-bit bounds and tnums.

Upon further inspection, I realized that, while the verifier bypasses
the above functions for mov32, it maintains a set of weaker invariants
that are still sufficient to verify the other ops. This commit replaces
my earlier invariants with this weaker invariant, and adds verification
for mov{32,64}.